+++ tensorflow_datasets/core/naming_test.py
@@ -1,5 +1,5 @@
 # coding=utf-8
-# Copyright 2019 The TensorFlow Datasets Authors.
+# Copyright 2020 The TensorFlow Datasets Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -63,6 +63,15 @@ class NamingTest(parameterized.TestCase, testing.TestCase):
   def test_filename_prefix_for_split(self, prefix, expected):
     split = splits.Split.TRAIN
     self.assertEqual(expected, naming.filename_prefix_for_split(prefix, split))
+
+  def test_filenames_for_dataset_split(self):
+    self.assertEqual([
+        "foo-train-00000-of-00002",
+        "foo-train-00001-of-00002",
+    ], naming.filenames_for_dataset_split(
+        dataset_name="foo",
+        split=splits.Split.TRAIN,
+        num_shards=2))
 
   def test_filepaths_for_dataset_split(self):
     self.assertEqual([