+++ tensorflow_datasets/image/dtd_test.py
@@ -1,5 +1,5 @@
 # coding=utf-8
-# Copyright 2019 The TensorFlow Datasets Authors.
+# Copyright 2020 The TensorFlow Datasets Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -30,10 +30,6 @@ class DtdTest(testing.DatasetBuilderTestCase):
       'train': 2,
       'validation': 1,
   }
-
-
-class DtdS3Test(DtdTest):
-  VERSION = 'experimental_latest'
 
 
 if __name__ == '__main__':