+++ tensorflow_datasets/image/svhn_test.py
@@ -1,5 +1,5 @@
 # coding=utf-8
-# Copyright 2019 The TensorFlow Datasets Authors.
+# Copyright 2020 The TensorFlow Datasets Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -34,10 +34,6 @@ class SvhnTest(testing.DatasetBuilderTestCase):
       "test": "test_32x32.mat",
       "extra": "extra_32x32.mat",
   }
-
-
-class SvhnS3Test(SvhnTest):
-  VERSION = "experimental_latest"
 
 
 if __name__ == "__main__":